Add the demo file
Whatever the game will be, the basic structure will always be the same. That is, loading surfaces and texts and receiving inputs in a loop. However to make it readable, and implement every fancy skill we need, there are still works to do. I have commented most places waiting to be improved. Please read.